♻️ refactor: service 코드를 인터페이스 - 구현체 형태로 리팩토링
