+++ src/test/java/com/grow/member_service/history/point/application/service/impl/PointCommandServiceImplTest.java
@@ -38,7 +38,8 @@ class PointCommandServiceImplTest {
 	@Mock PointHistoryRepository historyRepository;
 	@Mock ApplicationEventPublisher events;
 
-	@InjectMocks PointCommandServiceImpl sut;
+	@InjectMocks
+	PointCommandServiceImpl sut;
 
 	private Member newMember(Long id, int totalPoint) {
 		MemberProfile profile = new MemberProfile("u@test.com", "nick", null, Platform.KAKAO, "pid");
